backend/pyproject.toml
[project]
name = "mock-chatbot-backend"
version = "0.1.0"
description = "FastAPI backend for mock chatbot"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "stripe>=7.0.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
]